requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.3
requests>=2.31.0
python-dotenv>=1.0.1
pytest>=8.1.1
websockets>=12.0
sentence-transformers>=2.7.0
sqlalchemy>=2.0.29
dashscope>=1.14.0
python-dotenv>=1.0.1
